Updates for Numpy 2.0 compatibility on Windows, build, test and publish wheels

+++ skfmm/base_marcher.h
@@ -18,7 +18,7 @@ extern "C" {
 class baseMarcher
 {
 public:
-  baseMarcher(double *phi,      double *dx,  long *flag,
+  baseMarcher(double *phi,      double *dx,  long long *flag,
               double *distance, int ndim,    int *shape,
               bool self_test,   int order,   double narrow,
               int periodic);
@@ -94,7 +94,7 @@ protected:
   double          * distance_; // return value modified in place
   double          * phi_;
   double          * dx_;
-  long            * flag_;
+  long long       * flag_;
   int               error_;
   int               dim_;            // number of dimensions
   int               size_;           // flat size
